Add {To,From}RistrettoBytes
Edwards25519 is not a group of prime order --- it has order 8 * l, where l
is the (prime) order of the basepoint.  Mike Hamurg's Ristretto is a
convenient method to construct a group of order l from Edwards25519 as a
subquotient (quotient of a subgroup).

Half of the points of Edwards25519 are "in" the Ristretto group.  (It's
the even points.)  Every element of the Ristretto group has exactly four
points that represent it in the Edwards25519 curve. (Points are
equivalent iff they differ by a point of order less than or equal 4.)

This commit adds two functions.

ToRistrettoBytes() takes an even curvepoint and returns a binary
encoding of the corresponding tlement of the Ristretto group.

FromRistrettoBytes() takes a buffer and checks whether it is a binary
encoding of an element of the Ristretto group.  If it is, the function
returns a point that represents the encoded element of the Ristretto group.